+++ include/rudiments/private/modemclient.h
@@ -2,7 +2,7 @@
 		const char	*connectscript;
 		const char	*disconnectscript;
 		const char	*phonenumber;
-		unsigned int	retrywait;
-		unsigned int	retrycount;
+		unsigned long	retrywait;
+		unsigned long	retrycount;
 
 		void	modemclientClone(const modemclient &m);
